Result doesn't build yet.

llvm-svn: 53163

+++ llvm/include/llvm/Analysis/AliasSetTracker.h
@@ -19,9 +19,9 @@
 
 #include "llvm/Support/CallSite.h"
 #include "llvm/Support/Streams.h"
-#include "llvm/ADT/iterator"
-#include "llvm/ADT/hash_map"
-#include "llvm/ADT/ilist"
+#include "llvm/ADT/iterator.h"
+#include "llvm/ADT/hash_map.h"
+#include "llvm/ADT/ilist.h"
 
 namespace llvm {
 
@@ -232,7 +232,7 @@ private:
                   bool KnownMustAlias = false);
   void addCallSite(CallSite CS, AliasAnalysis &AA);
   void removeCallSite(CallSite CS) {
-    for (unsigned i = 0, e = CallSites.size(); i != e; ++i)
+    for (size_t i = 0, e = CallSites.size(); i != e; ++i)
       if (CallSites[i].getInstruction() == CS.getInstruction()) {
         CallSites[i] = CallSites.back();
         CallSites.pop_back();
